+++ src/JSystem/JKernel/JKRDecomp.cpp
@@ -43,7 +43,7 @@ void* JKRDecomp::run()
 		}
 
 		if (command->mCallback) {
-			(*command->mCallback)((u32)command);
+			(*command->mCallback)((u32)(uintptr_t)command);
 			continue;
 		}
